requirements.txt
beautifulsoup4==4.9.3
Flask==2.0.0
pyppeteer==0.2.5
requests-html==0.10.0
python-slugify==5.0.2